resume-screener/requirements.txt
faiss-cpu==1.9.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
tqdm==4.67.0
pyyaml==6.0.2
rank-bm25==0.2.2